etl_pipeline/requirements.txt
torch
torchvision
mlflow
tqdm
pyyaml
Pillow
fastapi
uvicorn
python-multiprocessing
onnx 
onnxruntime
